+++ django_fido/settings.py
@@ -3,7 +3,7 @@ from typing import List, Optional, cast
 
 from appsettings import (AppSettings, BooleanSetting, CallablePathSetting, FileSetting, NestedDictSetting,
                          NestedListSetting, PositiveIntegerSetting, Setting, StringSetting)
-from django.core.exceptions import ValidationError
+from django.core.exceptions import ImproperlyConfigured, ValidationError
 
 
 def timeout_validator(value):
@@ -36,6 +36,19 @@ class DjangoFidoSettings(AppSettings):
         crl_list=NestedListSetting(inner_setting=FileSetting(), default=[]),
     ), default=None)
     resident_key = BooleanSetting(default=False)
+    passwordless_auth = BooleanSetting(default=False)
+
+    @classmethod
+    def check(cls):
+        """Extend parent class check method to perform further project specific settings check."""
+        super(DjangoFidoSettings, cls).check()
+
+        # check passwordless settings
+        if cls.settings['passwordless_auth'].get_value() and not cls.settings['resident_key'].get_value():
+            raise ImproperlyConfigured("To use passwordless auth, RESIDENT_KEY settings must be set to True")
+
+        if cls.settings['passwordless_auth'].get_value() and cls.settings['two_step_auth'].get_value():
+            raise ImproperlyConfigured("To use passwordless auth, TWO_STEP_AUTH must be set to False")
 
     class Meta:
         """Meta class."""