+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog #
 ## Unreleased ##
  * Fix Authenticator.metadata if no identifier is present
+ * Added passwordless auth backend
 
 ## 0.37
  * Drop support for Django 1.11